requirements.txt
# Explainable Multimodal CNN-RNN for Chest X-Ray Diagnosis
# See PROJECT_PLAN.md for architecture details

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Medical imaging / CXR models
torchxrayvision>=1.2.0

# NLP / Transformers
transformers>=4.35.0
tokenizers>=0.15.0

# Explainability (XAI)
shap>=0.44.0
captum>=0.7.0
grad-cam>=1.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Data handling
h5py>=3.9.0
pydicom>=2.4.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Progress / logging
tqdm>=4.66.0
tensorboard>=2.15.0

# Web UI (optional)
streamlit>=1.29.0

# Jupyter (optional)
jupyter>=1.0.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0

# Linting (dev)
ruff>=0.1.0
